Replace example DAG stubs with operator stub
